==> 1_Merging_Data.ipynb <==
import pandas as pd

# File path
file_path = r"/content/customer_scorecard_input.xlsx"

# Read all 2 sheets
application_df = pd.read_excel(file_path, sheet_name="Application_Data")
behavioral_df = pd.read_excel(file_path, sheet_name="Behavioral_Data")

# Bureau data
file_path = r"/content/bureau_data.xlsx"
bureau_df = pd.read_excel(file_path, sheet_name="Bureau_Data")


# Merging them into one dataset using Customer_ID
merged_df = application_df.merge(behavioral_df, on="Customer_ID") \
                          .merge(bureau_df, on="Customer_ID")

# Quick check
print("Application Data shape:", application_df.shape)
print("Behavioral Data shape:", behavioral_df.shape)
print("Bureau Data shape:", bureau_df.shape)
print("Final merged shape:", merged_df.shape)

# Show first few rows
print(merged_df.head(10))

from google.colab import files

file_name = "merged_customer_data.xlsx"
merged_df.to_excel(file_name, index=False)

files.download(file_name)
